+++ tests/models/test_post.py
@@ -1,5 +1,5 @@
 from django.test import TestCase
-from blog.models import Post, Comment
+from blog.models import Post
 from django.contrib.auth import get_user_model
 from freezegun import freeze_time
 
@@ -19,14 +19,3 @@ class PostModelTest(TestCase):
         post = Post.objects.get(title='Test post')
         post.publish()
         self.assertEqual(post.published_date.strftime("%Y-%m-%d %H:%M:%S"), '2019-01-01 00:01:01')
-
-
-class CommentModelTest(TestCase):
-    def setUp(self):
-        user = get_user_model().objects.create(username='testuser')
-        Post.objects.create(title='Test post', author=user)
-
-    def test_a_comment_is_by_default_not_approved(self):
-        post = Post.objects.get(title='Test post')
-        comment = Comment.objects.create(post=post)
-        self.assertFalse(comment.approved)
